CARDS-2348: New question type: postal address Separated google API related functionality to a stand-alone component
