+++ modules/google-api/src/main/frontend/package.json
@@ -1,0 +1,44 @@
+{
+  "name": "cards-google-api",
+  "version": "0.1.0",
+  "description": "Google API key management",
+  "author": "UHN DATA",
+  "license": "Apache-2.0",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/data-team-uhn/cards.git",
+    "directory": "modules/data-entry/src/main/frontend"
+  },
+  "babel": {
+    "presets": [
+      "@babel/preset-env",
+      "@babel/preset-react"
+    ],
+    "plugins": [
+      ["@babel/plugin-proposal-class-properties"],
+      ["@babel/plugin-proposal-optional-chaining"]
+    ]
+  },
+  "devDependencies": {
+    "@babel/core": "7.21.0",
+    "@babel/plugin-proposal-class-properties": "7.18.6",
+    "@babel/plugin-proposal-optional-chaining": "7.11.0",
+    "@babel/preset-env": "7.21.4",
+    "@babel/preset-react": "7.18.6",
+    "babel-loader": "9.1.2",
+    "clean-webpack-plugin": "4.0.0",
+    "eslint-plugin-react": "7.32.0",
+    "eslint-plugin-auto-import": "0.1.1",
+    "eslint": "8.32.0",
+    "webpack": "5.77.0",
+    "webpack-cli": "4.4.0",
+    "webpack-assets-manifest": "5.1.0"
+  },
+  "dependencies": {
+    "@mui/icons-material": "5.11.0",
+    "@mui/material": "5.11.3",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-router-dom": "5.2.0"
+  }
+}
