+++ modules/google-api/src/main/frontend/webpack.config.js
@@ -1,0 +1,35 @@
+const { CleanWebpackPlugin } = require('clean-webpack-plugin');
+const WebpackAssetsManifest = require('webpack-assets-manifest');
+
+module_name = require("./package.json").name + ".";
+
+module.exports = {
+  mode: 'development',
+  entry: {
+    [module_name + 'GoogleApiKey']: { 'dependOn': ['cards-login.loginDialogue'], 'import': './src/googleApiKeyAdminPage.jsx' },
+    [module_name + 'googleIcon']: '@mui/icons-material/Google',
+  },
+  plugins: [
+    new CleanWebpackPlugin(),
+    new WebpackAssetsManifest({
+      output: "assets.json"
+    })
+  ],
+  module: {
+    rules: [
+      {
+        test: /\.(js|jsx)$/,
+        exclude: /node_modules/,
+        use: ['babel-loader']
+      }
+    ]
+  },
+  resolve: {
+    extensions: ['*', '.js', '.jsx']
+  },
+  output: {
+    path: __dirname + '/dist/SLING-INF/content/libs/cards/resources/',
+    publicPath: '/',
+    filename: '[name].[contenthash].js'
+  }
+};
