+++ modules/google-api/src/main/java/io/uhndata/cards/google/DefaultGoogleApiKeyManager.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package io.uhndata.cards;
+package io.uhndata.cards.google;
 
 import javax.jcr.Node;
