+++ modules/google-api/src/main/resources/SLING-INF/content/Extensions/AdminDashboard/AdminGoogleApiKey.json
@@ -3,7 +3,7 @@
   "cards:extensionPointId": "cards/coreUI/admindashboard/entry",
   "cards:extensionName": "Google Api Key",
   "cards:targetURL": "/content.html/admin/GoogleApiKey",
-  "cards:icon": "asset:cards-dataentry.googleIcon.js",
+  "cards:icon": "asset:cards-google-api.googleIcon.js",
   "cards:defaultOrder": 98,
   "cards:hint": "Use Google API services"
 }
