+++ modules/google-api/src/main/resources/SLING-INF/content/libs/cards/resources/assetDependencies.json
@@ -1,0 +1,3 @@
+{
+  "cards-google-api.GoogleApiKey.js": ["asset:cards-login.loginDialogue.js"]
+}
