+++ modules/pom.xml
@@ -74,5 +74,6 @@
     <module>slack-notifications</module>
     <module>embedded-view</module>
     <module>error-tracking</module>
+    <module>google-api</module>
   </modules>
 </project>
